refactor(issue-463): Simplify and abstract Storybook stories
- Extract shared mock data into helper files
- Remove trivial stories, keep only meaningful variations
- Point breadcrumbs testing to dedicated test harness
- Create reusable mock state creators
